Unified platforms tab with Kubernetes support
