+++ ui/package.json
@@ -25,6 +25,7 @@
     "babel-loader": "^8.0.5",
     "copy-webpack-plugin": "^4.6.0",
     "css-loader": "^5.2.6",
+    "file-loader": "^6.2.0",
     "html-webpack-plugin": "^3.2.0",
     "terser-webpack-plugin": "4.x",
     "webpack": "^4.29.3",
