--- ui/src/components/pages/tenancy/kubernetes-clusters/index.js
@@ -1,2 +1,0 @@
-import { TenancyKubernetesClustersPanel } from './panel';
-export { TenancyKubernetesClustersPanel };
