+++ ui/src/components/pages/tenancy/platforms/clusters/index.js
@@ -1,0 +1,2 @@
+import { ClusterCard } from './card';
+export { ClusterCard };
