+++ ui/src/components/pages/tenancy/platforms/kubernetes/index.js
@@ -1,0 +1,2 @@
+import { KubernetesCard } from './card';
+export { KubernetesCard };
