+++ ui/src/components/pages/tenancy/platforms/kubernetes/upgrade-modal.js
@@ -1,23 +1,28 @@
-import React, { useState } from 'react';
+import React, { useEffect, useState } from 'react';
 
 import Alert from 'react-bootstrap/Alert';
 import Button from 'react-bootstrap/Button';
-import DropdownItem from 'react-bootstrap/DropdownItem';
 import Modal from 'react-bootstrap/Modal';
 
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { faArrowAltCircleUp } from '@fortawesome/free-solid-svg-icons';
+import {
+    faArrowAltCircleUp,
+    faShieldAlt,
+    faSyncAlt
+} from '@fortawesome/free-solid-svg-icons';
 
-import { Form, Field } from '../../../utils';
+import { Form, Field } from '../../../../utils';
 
-import { KubernetesClusterTemplateSelectControl } from '../resource-utils';
+import { KubernetesClusterTemplateSelectControl } from '../../resource-utils';
 
 
-export const UpgradeKubernetesClusterMenuItem = ({
+export const UpgradeKubernetesClusterButton = ({
     kubernetesCluster,
     kubernetesClusterActions,
     kubernetesClusterTemplates,
-    kubernetesClusterTemplateActions
+    kubernetesClusterTemplateActions,
+    disabled,
+    ...props
 }) => {
     const [visible, setVisible] = useState(false);
     const open = () => setVisible(true);
@@ -26,27 +31,30 @@ export const UpgradeKubernetesClusterMenuItem = ({
     const [template, setTemplate] = useState(kubernetesCluster.template.id);
     const reset = () => setTemplate(kubernetesCluster.template.id);
 
-    const handleCancel = () => {
-        reset();
-        close();
-    };
+    const [inFlight, setInFlight] = useState(false);
+    useEffect(
+        () => { if( inFlight && !kubernetesCluster.updating ) setInFlight(false); },
+        [!!kubernetesCluster.updating]
+    );
 
     const handleSubmit = (evt) => {
         evt.preventDefault();
-        reset();
-        close();
         kubernetesClusterActions.update({ template }, true);
+        setInFlight(true);
+        close();
     };
 
     return (
         <>
-            <DropdownItem
-                onSelect={open}
-                disabled={kubernetesCluster.status.endsWith("ing")}
-            >
-                Upgrade cluster
-            </DropdownItem>
-            <Modal size="lg" backdrop="static" onHide={handleCancel} show={visible}>
+            <Button {...props} variant="warning" disabled={disabled} onClick={open}>
+                <FontAwesomeIcon
+                    icon={inFlight ? faSyncAlt : faShieldAlt}
+                    spin={inFlight}
+                    className="me-2"
+                />
+                {inFlight ? 'Upgrading...' : 'Upgrade'}
+            </Button>
+            <Modal size="lg" backdrop="static" onHide={close} onExited={reset} show={visible}>
                 <Modal.Header closeButton>
                     <Modal.Title>Upgrade Kubernetes cluster {kubernetesCluster.name}</Modal.Title>
                 </Modal.Header>
